+++ src/main/java/by/bulaukin/contactlist/services/ContactsServices.java
@@ -10,5 +10,5 @@ public interface ContactsServices {
     Contact findById(long id);
     Contact save(Contact contact);
     Contact update(Contact contact);
-    void deleteBYId(int id);
+    void deleteBYId(long id);
 }
